requirements.txt
requests
torch
transformers
pandas
numpy
python-dotenv
peft
pytz
